Hua_Dian_QA/requirements.txt
# For GPU acceleration, please install the correct PyTorch version for your system first.
# Visit https://pytorch.org/get-started/locally/
# Example for CUDA 12.1:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

langchain
langchain-community
langchain-core
langchain-ollama
langchain-huggingface
sentence-transformers
lancedb
pylance
pyarrow
pytorch-lightning
tqdm
PyMuPDF
gradio
lxml
python-docx
unstructured
fastapi
uvicorn
langchain-ollama